# Ring-buffer based runtime tracing (eventring)