--- a/Makefile.common
+++ b/Makefile.common
@@ -93,7 +93,7 @@ endif # ifeq "$(wildcard $(ROOTDIR)/flexlink.opt$(EXE))" ""
 endif # ifeq "$(BOOTSTRAPPING_FLEXDLL)" "false"
 
 # List of other libraries
-ALL_OTHERLIBS = dynlink str systhreads unix
+ALL_OTHERLIBS = dynlink str systhreads unix runtime_events
 
 # The following variable defines flags to be passed to the C preprocessor
 # when compiling C files for the native runtime.