--- a/otherlibs/runtime_events/.depend
+++ b/otherlibs/runtime_events/.depend
@@ -1,0 +1,5 @@
+runtime_events.cmo : \
+    runtime_events.cmi
+runtime_events.cmx : \
+    runtime_events.cmi
+runtime_events.cmi :