--- a/runtime/array.c
+++ b/runtime/array.c
@@ -23,7 +23,7 @@
 #include "caml/misc.h"
 #include "caml/mlvalues.h"
 #include "caml/signals.h"
-#include "caml/eventlog.h"
+#include "caml/runtime_events.h"
 
 static const mlsize_t mlsize_t_max = -1;