--- a/testsuite/tests/lib-runtime-events/test.reference
+++ b/testsuite/tests/lib-runtime-events/test.reference
@@ -1,0 +1,2 @@
+minors: 9, majors: 3
+minors: 9, majors: 3