## Summary

Eventring is a runtime tracing system designed for continuous monitoring of OCaml applications. It intends to replace the existing file-logging based eventlog in the runtime. It is very low overhead - negligible when not in use and ~0.5% when enabled. It utilises per-domain memory-mapped ring buffers that enable consumption of tracing events both in and out of process. At present only runtime events (the same set as eventlog) are added to the buffers but the design is intended to also eventually support user events as well as integrate with statmemprof.

# Design

# Context

The OCaml runtime currently provides an event logging system "eventlog" which can emit a range of runtime and GC events in the Common Trace Format (CTF) to a log file. It is currently only enabled in the instrumented runtime. The file-based logging also makes using eventlog for long running applications very difficult.

# Goal

The goal of this PR is to replace eventlog's implementation with a minimal implementation of a more flexible transport for runtime and GC events. It intends to enable consuming of these events from OCaml code both inside and outside the runtime being monitored.

This PR only intends to deal with runtime and GC events, not those originating from user code. There is follow-up work on porting eventlog-tools, documentation and making the Windows support more robust (and potentially more native). We hold off on those if there are design changes required at this stage and also because the PR is already fairly large.

# Design

## Probes

This PR retains the existing probes on trunk but moves most in to the default runtime. A few of the probes on hot paths (such as free list allocation) are left in the instrumentated runtime. The dividing line is essentially where the work performed during a particular event is very high compared to the (low) cost of emitting an event.

## Transport

The on-disk log file in eventlog is replaced with per-domain memory-mapped ring buffers. 

The ring buffer is structured as a flight recorder, overwriting old data when is insufficient space to write new events. This enables users to potentially only read the ring when some anomalous event occurs or a performance target is missed. No coordination is needed with consumers who read the events - they detect races with the producer and discard events when that happens.

The producer code is contained here . By default an .eventring file is created in the current directory (overridable by setting OCAML_EVENTRING_DIR). This file contains a ring buffer for each possible domain (Max_domains). It is laid out in a structure that enables sparsity. On-disk (or in-memory) footprint is proportional to the max number of concurrent domains the process has ever run.

#### On-disk structure
```
----------------------------------------------------------------
| File header (version, offsets, etc..)                        |
----------------------------------------------------------------
| Ring 0..Max_domains metadata                                 |
| (head and tail indexes, one per cache line)                  |
----------------------------------------------------------------
| Ring 0..Max_domains data                                     |
| (actual ring data, default 2^16 words = 512k bytes)          |
----------------------------------------------------------------
```

## API

The consumer APIs are cursor-based and polled. Users create a cursor with a path and pid, and if successfully created can then poll the cursor with a set of callbacks to call on each event they are interested in.

The OCaml API is available in [Eventring.mli](https://github.com/sadiqj/ocaml/blob/eventring-pr/stdlib/eventring.mli), the C API is available in [eventring.h](https://github.com/sadiqj/ocaml/blob/eventring-pr/runtime/caml/eventring.h).


Consumer code is contained in [eventring_consumer.c](https://github.com/sadiqj/ocaml-multicore/blob/eventring_multicore/runtime/eventring_consumer.c). This contains both the implementations for the OCaml API and C API (the former being a thin layer over the latter).

## Examples

### Emitting minor and major slice GC pause times to stdout

Here's a simple consumer that prints the times of minor and major slices. It omits any support for multiple domains, error handling or lost events:

```let major_started = ref Int64.zero
let minor_started = ref Int64.zero

let runtime_begin _ ts phase =
    match phase with
    | Eventring.EV_MAJOR_SLICE ->
        begin
          major_started := Eventring.Timestamp.to_int64 ts
        end
    | Eventring.EV_MINOR ->
        begin
          minor_started := Eventring.Timestamp.to_int64 ts
        end
    | _ -> ()


let ts_to_duration_ns ts0 ts1 =
  Int64.sub (Eventring.Timestamp.to_int64 ts0) ts1

let runtime_end _ ts phase =
    match phase with
    | Eventring.EV_MAJOR_SLICE ->
        Printf.printf "Major slice: %Ldns\n"
          (ts_to_duration_ns ts !major_started)
    | Eventring.EV_MINOR ->
        Printf.printf "Minor: %Ldns\n"
          (ts_to_duration_ns ts !minor_started)
    | _ -> ()

let () =
    let list_ref = ref [] in
    let cursor = Eventring.create_cursor None in
    let callbacks = Eventring.Callbacks.create ~runtime_begin ~runtime_end ()
    in
    while true do
        for _ = 1 to 100 do
            list_ref := [];
            for _ = 1 to 10 do
                list_ref := (Sys.opaque_identity(ref 42)) :: !list_ref
            done;
            Gc.full_major ();
        done;
        ignore(Eventring.read_poll cursor callbacks None);
        Unix.sleep 1
    done
```

When compiled and run with this PR's branch you will get an output like:

```
Minor: 1496ns
Minor: 1341ns
Minor: 1308ns
Major slice: 17605ns
Minor: 1435ns
Minor: 2986ns
Minor: 2688ns
Major slice: 26577ns
Minor: 2263ns
Minor: 1098ns
Minor: 973ns
Major slice: 17719ns
...
```

Code such as the above could be spawned in a separate domain and send metrics data to an external service (such as graphite or prometheus) over a socket.

### Chrome tracing

[This](https://github.com/sadiqj/traceevents_lib/blob/main/src/traceevents_lib.ml) is a simplified example of how Eventring can be used to log out runtime data in a format that can be simply ingested by Chrome's trace viewer. Running the above project with `OCAML_EVENTRING_START=1 _build/default/src/traceevents_lib.exe` and then loading the resultant json file in to Chrome's tracing viewer will give something like the following:

![Example of chrome tracing outpu](https://user-images.githubusercontent.com/631805/146253984-ea161c17-2e7b-4bce-855a-00678c91eee3.png)

### Tests

The [tests/eventring](https://github.com/sadiqj/ocaml-multicore/tree/eventring_multicore/testsuite/tests/eventring) directory contains some other exampel of using the different APIs.

# Benchmarks

# Timing

For all macro benchmarks on sandmark that run on trunk, the performance impact when normalised to trunk (f22a84b4fdaafa92ed2a6a783ec8f8d249baf78) averaged over five runs is:

| Variant | Geomean |
|-----------|---------------|
| Eventring disabled | 0.998507 |
| Eventring enabled  | 0.997360 |

<1 indicates eventring's branches were marginally faster than trunk.

The graph below shows the normalised (to trunk) runtime for long-running (>4s) macro benchmarks:

![image](https://user-images.githubusercontent.com/631805/152162426-c3e2e90f-1ffe-4a17-9537-c0b6d10cf15d.png)

All macro benchmarks:

![image](https://user-images.githubusercontent.com/631805/152162347-4b6f1e8a-e2b4-4e12-98ad-352477f35233.png)

You can find the raw benchmarking data here: https://github.com/sadiqj/sandmark_benchmark_data/tree/main/eventring_pr

# Instructions

Below is a run of sequential sandmark using perf stat to count the number of retired instructions:

![image](https://user-images.githubusercontent.com/631805/152160368-9cae6e4e-5ae2-407b-8dd4-f79ba729a055.png)

| Variant | Geomean |
|-----------|---------------|
| Eventring disabled | 1.000133 |
| Eventring enabled  | 1.000913 |

Here eventring resulted an 0.01% increase in instructions when not enabled and 0.09% when enabled.

For each benchmark below there are two rows, this PR without eventring enabled and this PR with it enabled. Instructions increase is the percentage increase in instructions for this PR over trunk e.g for the first row `LU_decomposition` this PR leads to an increase of 7 thousandths of a percent (2.25M extra instructions on the 32.28B trunk does).

| name | instructions | variant | ninstructions |
| -- | -- | -- | -- |
LU_decomposition 1024 | 32.284B | disabled | 0.007% |
LU_decomposition 1024 | 32.338B | enabled | 0.175%
alt-ergo benchmarkalt-ergfill.why | 10.726B | disabled | 0.004%
alt-ergo benchmarkalt-ergfill.why | 10.729B | enabled | 0.035%
alt-ergo benchmarkalt-ergyyll.why | 80.394B | disabled | 0.005%
alt-ergo benchmarkalt-ergyyll.why | 80.449B | enabled | 0.074%
bdd 26 | 23.379B | disabled | 0.005%
bdd 26 | 23.410B | enabled | 0.139%
binarytrees5 21 | 64.412B | disabled | 0.005%
binarytrees5 21 | 64.438B | enabled | 0.045%
chameneos_redux_lwt 600000 | 8.287B | disabled | 0.007%
chameneos_redux_lwt 600000 | 8.302B | enabled | 0.196%
crout_decomposition | 10.837B | disabled | 0.004%
crout_decomposition | 10.851B | enabled | 0.139%
cubicle benchmarkcubiclgerman_pfs.cub | 1011.065B | disabled | 0.009%
cubicle benchmarkcubiclgerman_pfs.cub | 1011.628B | enabled | 0.064%
cubicle benchmarkcubiclszymanski_at.cub | 2722.539B | disabled | 0.006%
cubicle benchmarkcubiclszymanski_at.cub | 2725.030B | enabled | 0.098%
durand_kerner_aberth | 1.054B | disabled | 0.008%
durand_kerner_aberth | 1.057B | enabled | 0.268%
evolutionary_algorithm 10000 10000 | 253.398B | disabled | 0.002%
evolutionary_algorithm 10000 10000 | 253.406B | enabled | 0.005%
fannkuchredux 12 | 587.338B | disabled | -0.001%
fannkuchredux 12 | 587.338B | enabled | -0.000%
fannkuchredux2 12 | 590.380B | disabled | 0.001%
fannkuchredux2 12 | 590.380B | enabled | 0.001%
fasta3 25_000_000 | 21.914B | disabled | 0.002%
fasta3 25_000_000 | 21.926B | enabled | 0.056%
fasta6 25_000_000 | 17.562B | disabled | -0.000%
fasta6 25_000_000 | 17.562B | enabled | -0.001%
fft | 19.972B | disabled | 0.005%
fft | 19.982B | enabled | 0.059%
floyd_warshall 512 | 6.032B | disabled | 0.002%
floyd_warshall 512 | 6.031B | enabled | -0.008%
game_of_life 256 | 64.563B | disabled | 0.000%
game_of_life 256 | 64.563B | enabled | 0.001%
grammatrix | 447.579B | disabled | -0.000%
grammatrix | 447.581B | enabled | 0.000%
kb | 16.056B | disabled | 0.003%
kb | 16.071B | enabled | 0.097%
kb_no_exc | 14.357B | disabled | 0.270%
kb_no_exc | 14.364B | enabled | 0.318%
knucleotide | 290.074B | disabled | 0.000%
knucleotide | 290.075B | enabled | 0.001%
knucleotide3 | 304.365B | disabled | 0.000%
knucleotide3 | 304.365B | enabled | 0.000%
levinson_durbin | 24.481B | disabled | 0.008%
levinson_durbin | 24.516B | enabled | 0.150%
main | 64.580B | disabled | 0.002%
main | 64.623B | enabled | 0.069%
mandelbrot6 16_000 | 185.860B | disabled | -0.000%
mandelbrot6 16_000 | 185.860B | enabled | 0.000%
matrix_multiplication 1024 | 32.692B | disabled | -0.000%
matrix_multiplication 1024 | 32.692B | enabled | 0.001%
menhir -v --table benchmarkmenhisysver.mly | 423.416B | disabled | 0.007%
menhir -v --table benchmarkmenhisysver.mly | 423.641B | enabled | 0.060%
menhir -v -t benchmarkmenhikeywords.mly benchm... | 37.790B | disabled | 0.008%
menhir -v -t benchmarkmenhikeywords.mly benchm... | 37.820B | enabled | 0.087%
menhir benchmarkmenhiocaml.mly --list-errors -... | 1154.743B | disabled | 0.000%
menhir benchmarkmenhiocaml.mly --list-errors -... | 1154.808B | enabled | 0.006%
minilight-ocaml benchmarkminilighroomfront.ml.txt | 80.788B | disabled | 0.015%
minilight-ocaml benchmarkminilighroomfront.ml.txt | 81.023B | enabled | 0.307%
naive_multilayer | 33.856B | disabled | 0.117%
naive_multilayer | 33.911B | enabled | 0.281%
nbody 50_000_000 | 40.758B | disabled | -0.000%
nbody 50_000_000 | 40.759B | enabled | 0.001%
pidigits5 10_000 | 11.882B | disabled | 0.044%
pidigits5 10_000 | 11.890B | enabled | 0.116%
qr_decomposition | 19.412B | disabled | 0.007%
qr_decomposition | 19.455B | enabled | 0.228%
quicksort 4000000 | 17.125B | disabled | -0.000%
quicksort 4000000 | 17.125B | enabled | 0.001%
regexredux2 | 93.000B | disabled | -0.002%
regexredux2 | 93.003B | enabled | 0.002%
revcomp2 | 20.703B | disabled | 0.002%
revcomp2 | 20.705B | enabled | 0.011%
sequence_cps 10000 | 14.535B | disabled | 0.011%
sequence_cps 10000 | 14.576B | enabled | 0.294%
setrip -enc -rseed 1067894368 | 2.454B | disabled | 0.002%
setrip -enc -rseed 1067894368 | 2.456B | enabled | 0.075%
spectralnorm2 5_500 | 59.332B | disabled | 0.003%
spectralnorm2 5_500 | 59.393B | enabled | 0.106%
test_decompress 64 524_288 | 28.016B | disabled | -0.013%
test_decompress 64 524_288 | 28.028B | enabled | 0.029%
test_lwt 200 | 136.513B | disabled | 0.001%
test_lwt 200 | 136.537B | enabled | 0.019%
thread_ring_lwt_mvar 20_000 | 20.404B | disabled | 0.007%
thread_ring_lwt_mvar 20_000 | 20.434B | enabled | 0.154%
thread_ring_lwt_stream 20_000 | 63.936B | disabled | 0.007%
thread_ring_lwt_stream 20_000 | 64.001B | enabled | 0.108%
ydump -c sample.json | 4.046B | disabled | -0.011%
ydump -c sample.json | 4.047B | enabled | 0.017%
zarith_pi 5000 | 3.006B | disabled | 0.063%
zarith_pi 5000 | 3.017B | enabled | 0.413%