@yminsky 2022-01-29
Love to see more work in the tracing space!

I wonder if this has any relation to the tracing library we released recently: https://github.com/janestreet/tracing

Or to our work on leveraging Intel's processor trace:  https://blog.janestreet.com/magic-trace/

Or to the work we've done adding probes to OCaml. Not sure if this is the best link for that: https://github.com/gretay-js/probes