@sadiqj 2022-01-29
> I am not sure using geomean is a good ground for performance impact argument. Looking at the chart some test cases became faster with eventring added, how come? And redux has 10% impact, is that right?

I should have been more specific above the table. The numbers are the average of the _normalised_ times against trunk (i.e normalised before averaging rather than after). Geometric mean is the right thing to do in this case: http://www.cse.unsw.edu.au/~cs9242/11/papers/Fleming_Wallace_86.pdf

Yes, some of the benchmarks show a speedup. This PR only adds instructions so you would think it would always make things slower but there's a couple of sources of noise. 

First is just the machine itself. While the benchmarking machine has steps to reduce variability (https://github.com/ocaml-bench/ocaml_bench_scripts/#notes-on-hardware-and-os-settings-for-linux-benchmarking) there is still some. I have a batch run of a number of iterations of the macro benchmarking suite going to quantify what the run-to-run variability is.

Second source is actually just small changes can have knock-on effects. e.g adding a probe could change the alignment of a later loop an lead to a significant slowdown. http://eecs.northwestern.edu/~robby/courses/322-2013-spring/mytkowicz-wrong-data.pdf has some depressing examples. Also @ctk21 has good notes on his investigations https://github.com/ocaml-bench/notes/blob/master/apr19.md#what-we-learnt .

After this large batch run finishes I'll also kick off a run that only uses instruction counts.

There's also benchmarks on parallel programs that need to be done. I'm waiting on this in-case we need to revisit the design or implementation significantly.