@sadiqj 2022-01-29
> Love to see more work in the tracing space!

Thanks for taking a look.

The aim for this PR is to only provide a facility for consuming events generated by the runtime's subsystems (mostly the GC but also some lifecycle events). This is intentional, I'm aware there's already some good observability tooling. Tracing, probes and memprof/memtrace being able to make use of eventring would be really powerful and I've already had a couple of discussions with JS folk on what some of that might look like. `Eio` also has an eye on replacing their tracing implementation with something atop of eventring.

It would be good to identify if there's anything in this design that precludes this from happening. Interfaces for plugging in a tracing or a probe library would be follow-up work requiring some community coordination.

Alternatively we might find we can't cover enough common usecases and eventring only functions as a way of transporting runtime events somewhere else. That's still better than the situation at the moment but it would feel like a missed opportunity.