@sadiqj 2022-01-31
I found the issue with `naive_multilayer`. It very frequently calls `caml_make_vect` which ultimately triggers a `handle_gc_interrupt` on each invocation and this causes an `EV_INTERRUPT_GC` to be emitted - even if there's nothing to do. Usually we only end up on this code path when there is something non-trivial to do (and so the relative cost of checking whether eventring is enabled and emitting the event is very low).

I was considering moving this event to the instrumented runtime but I'm actually not convinced it's that useful even there so will remove it and kick off another benchmarking run.