@sadiqj 2022-02-02
I've rerun and updated the benchmarking results after removing `EV_INTERRUPT_GC`.