@gasche 2022-02-02
I looked at the benchmarks and I'm a bit confused (as often).

The "retired instructions" measures look like we expect: at most there is 0.4% increase in instruction count when eventring is enabled, nothing to worry about, and when it is disabled it's mostly lost in the noise except for a couple outliers. (What's going on with these outliers? I wonder, but given that the difference is so small it's okay anyway.)

The "time" benchmarks exhibit wide noise, possibly due to code-placement issues: there are 10% changes in performance, either up or down, with no clear pattern or obvious relation to the change itself. This is not unexpected, and we shouldn't be surprised anymore. What I *do* find surprising is that there are a handful of outliers (one microbenchmark and two-four macrobenchmarks; really most cases where something "interesting" seems to happen) where the eventring-disabled version is *noticeably slower* than the eventring-enable solution. This is weird! In particular, both versions use the same code layout (right?), so the weird difference comes from something else. (And there are essentially no programs where the difference goes in the other direction, where eventring-enabled program is noticeably slower than the eventring-disabled program.)

Am I reading the graphs wrong (I expect that score >1 means "slower than the baseline", and <1 means "faster than the baseline".)? Could you have switched the two labels by mistake?