@sadiqj 2022-02-02
> I looked at the benchmarks and I'm a bit confused (as often).

Thanks for taking a look @gasche . Your comment did get me thinking and I realised that the `eventring-pr` and `eventring-enabled` branches could result in different runtimes. To better fit sandmark's workflow the `eventring-enabled` branch is `eventring-pr` with a one-liner commit atop that always starts eventring, regardless of whether the environment variable is passed.

So instead I kicked off another benchmarking run, this time with the same runtime but only using environment variables to turn eventring on. There's still benchmarks where eventring _enabled_ takes less time. 

I picked knucleotide3 and ran the binary (taskset on an isolated core) and averaged over 6 runs of each there's still a ~3.7% performance advantage to having eventring enabled. 

I wondered whether just having an extra environment variable the same size as `OCAML_EVENTRING_START` was enough but no dice, it performs the same as not enabled.