@gasche 2022-02-08
Personally I'm not so worried about the benchmarks: there is nothing in them that suggests that the changes does add noticeable overhead (we are discussing strange noise effects, which suggests that there is nothing else to discuss). And if we were to find noticeable performance impact later, we have the easy option of moving more probes from the normal runtime to the instrumented runtime. I would  consider this point settled.

But if we want to make a decision on the PR, we need:
1. decisions on whether we like the new design better than the previous Eventlog design
2. and a review of the implementation

So far we haven't had those. Point (1) sounds like an easy sell ("it comes from the same group, we said yes last time, they say it's even better this time, so what's not to like?"), and we have a caml-devel meeting shortly where you can always poll to force a decision. Point (2) is the bottleneck, as always.